frontend/backend-ml/requirements.txt
# Backend ML Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pytesseract>=0.3.10
Pillow>=10.0.0
PyPDF2>=3.0.0
pdf2image>=1.16.0
scikit-learn>=1.3.0
joblib>=1.3.0
numpy>=1.26.0
pandas>=2.1.0
python-dotenv>=1.0.0
